feat: add node click feature
